Set osMock in a single dedicated place
