+++ internal/statemachine/pack_test.go
@@ -2,7 +2,6 @@ package statemachine
 
 import (
 	"fmt"
-	"io/fs"
 	"os"
 	"os/exec"
 	"path/filepath"
@@ -14,50 +13,6 @@ import (
 	"github.com/canonical/ubuntu-image/internal/commands"
 	"github.com/canonical/ubuntu-image/internal/helper"
 )
-
-type osMockConf struct {
-	osutilCopySpecialFileThreshold uint
-	ReadDirThreshold               uint
-	RemoveThreshold                uint
-}
-
-type osMock struct {
-	conf                            *osMockConf
-	beforeOsutilCopySpecialFileFail uint
-	beforeReadDirFail               uint
-	beforeRemoveFail                uint
-}
-
-func (o *osMock) CopySpecialFile(path, dest string) error {
-	if o.beforeOsutilCopySpecialFileFail >= o.conf.osutilCopySpecialFileThreshold {
-		return fmt.Errorf("CopySpecialFile fail")
-	}
-	o.beforeOsutilCopySpecialFileFail++
-
-	return nil
-}
-
-func (o *osMock) ReadDir(name string) ([]fs.DirEntry, error) {
-	if o.beforeReadDirFail >= o.conf.ReadDirThreshold {
-		return nil, fmt.Errorf("ReadDir fail")
-	}
-	o.beforeReadDirFail++
-
-	return []fs.DirEntry{}, nil
-}
-
-func (o *osMock) Remove(name string) error {
-	if o.beforeRemoveFail >= o.conf.RemoveThreshold {
-		return fmt.Errorf("Remove fail")
-	}
-	o.beforeRemoveFail++
-
-	return nil
-}
-
-func NewOSMock(conf *osMockConf) *osMock {
-	return &osMock{conf: conf}
-}
 
 func TestPack_Setup(t *testing.T) {
 	t.Run("test_classic_setup", func(t *testing.T) {
